- Animate numbers in home statistics progressbars - rearranged icons

+++ app/src/main/java/host/stjin/anonaddy/utils/NumberUtils.kt
@@ -4,7 +4,7 @@ import java.math.RoundingMode
 
 object NumberUtils {
 
-    fun roundOffDecimal(number: Double): Double {
-        return number.toBigDecimal().setScale(1, RoundingMode.FLOOR).toDouble()
+    fun roundOffDecimal(number: Double): Float {
+        return number.toBigDecimal().setScale(1, RoundingMode.FLOOR).toFloat()
     }
 }
